+++ ressources/documentation-theme/jsdoc/config.js
@@ -4,7 +4,7 @@
 const fss = require("@absolunet/fss");
 
 const { source, destination } = JSON.parse(process.env.__ABSOLUNET_MANAGER_JSDOC_CONFIG__);
-const readme = fss.realpath(`${source}/../readme.md`);
+const readme = fss.realpath(`${source}/../README.md`);
 
 // eslint-disable-next-line no-undef
 module.exports = {