+++ src/managers/SingleManager.js
@@ -4,7 +4,6 @@
 import fss from "@absolunet/fss";
 import __ from "@absolunet/private-registry";
 import { terminal } from "@absolunet/terminal";
-import chalk from "chalk";
 import builder from "../helpers/builder.js";
 import documenter from "../helpers/documenter.js";
 import fixer from "../helpers/fixer.js";
@@ -21,7 +20,7 @@ class SingleManager extends AbstractManager {
 	/**
 	 * @inheritdoc
 	 */
-	get version() {
+	get currentVersion() {
 		if (fss.exists(paths.package.config)) {
 			const { version } = fss.readJson(paths.package.config);
 
@@ -29,22 +28,6 @@ class SingleManager extends AbstractManager {
 		}
 
 		return null;
-	}
-
-	/**
-	 * @inheritdoc
-	 */
-	install(options) {
-		return super.install(options, async () => {
-			// Symlink if self-reference
-			const config = fss.readJson(paths.package.config);
-			if (Object.keys(config.devDependencies).includes(config.name)) {
-				const dependenciesPath = `${paths.package.root}/node_modules/${config.name}`;
-				fss.remove(dependenciesPath);
-				fss.symlink(paths.package.root, dependenciesPath);
-				terminal.print(`Symlink self-reference dependency`).spacer();
-			}
-		});
 	}
 
 	/**
@@ -93,17 +76,12 @@ class SingleManager extends AbstractManager {
 	/**
 	 * @inheritdoc
 	 */
-	prepare(options) {
-		return super.prepare(options, async () => {
-			// Update version if self-reference
-			const config = fss.readJson(paths.package.config);
-			if (Object.keys(config.devDependencies).includes(config.name)) {
-				config.devDependencies[config.name] = config.version;
-				fss.writeJson(paths.package.config, config, { space: 2 });
-				terminal
-					.print(`Update self-reference version in ${chalk.underline(util.relativizePath(paths.package.config))}`)
-					.spacer();
-			}
+	version(options) {
+		return super.version(options, async () => {
+			const version = util.incrementVersion(this.currentVersion);
+
+			// Update version
+			terminal.process.run(`npm version ${version} --no-git-tag-version`);
 		});
 	}
 }